Simplified as_pyarray(), refactor find_intersectin
bind11_utils were uselessly bloated for the sake of creating a base
sequence directly in the heap. However, thanks to move semantics, the
cost of moving a vector is almost zero (its data is always in heap).
Therefore the user can create a normal vector and simply call
as_pyarray(&&) routine move and create the capsule

find_intersection was largely refactored to accept any iterator. In the
new design we use unary predicates for teh selection, resulting in a very
flexible API and way cleaner code

Python tests were added to the Cmake tests

Documentation...

Change-Id: Icf15f28a8ba744e02257d8f34243cfd89e5caf70