Open/create memory maps without ever moving the tree object
When creating/using Rtree from memory maps we were moving the object, created in the mem-map, to "itself", potentially in the heap. While elegant, this could lead to corruption in case the program aborted for any reason.

In this patch we avoid such complexities and instead design this manager object as a holder type for pybind11, with API compat with a managed pointer (similar to unique_ptr), so that it can also be friendly for direct uses.