Fix/syn gids serialize
**Context**

This MR addresses SPIND-133, in which the synapse source and dest gids were not correct after save-restoring.

**Scope**

Up until synapse indexes, all the id information was encoded in a single field (the `id`). However, synapses now
bring two additional ids as payload.

In this MR The serialization of the id field became generic to use the base class serialization. In the case of Synapses, `SynapseId` implements its own `serialize` method which considers also `post_gid_` and `pre_gid_`;