Improve handling of MPI startup executable.
On BB5 we observe that MVAPICH2 doesn't come with `mpirun` or `mpiexec`. While HPE MPI does offer a `mpiexec` but it's not standard compliant, because it uses `-np` instead of `-n` to specify the number of processes.

The truce negotiated in this MR is that we'll first check if `srun` exists. If it does: use it. If not we use whatever `FindMPI` decided is the executable, and rely on it to be implemented in a standard compliant manner.

If this fails, both the startup executable and the flag for specifying the number of processes, i.e. `-n` vs. `-np`, can be passed as arguments to CMake.

This also contains some generic clean up of the CI, by removing the never used feature to run Python tests through `ctest`.