Improve the random tree tests.
This MR consists of three parts.

The first adds the overloads of `operator<<(std::ostream& os, const Box3D& b)` to the namespace `boost::geometry::model`. The reason for this change is argument dependent lookup. In particular the problem arises because `Box3D` is an alias to `boost::geometry::model::box`. Any code we write within the namespace `spatial_index` would previously work and with `spatial_index` a `Box3D` is printable. However, in units tests we would like to use `BOOST_CHECK_MESSAGE(cond, "box = " << box)`. This is different: it will look in some boost unit testing namespace, the global namespace. Then it'll try ADL which means check in `std::` and `boost::geometry::model::`. Previously, it would fail; to compile.

The second part is to add constructors to enable the following:
```
   auto v = std::vector<Sphere>();
   v.emplace_back(centroid, radius);
```

In the third part we now check on a tree containing sphere, cylinder and box-like elements, with query shapes that are boxes, spheres and cylinders that

  - find_intersecting
  - count_intersecting
  - is_intersecting

are internally consistent. Additionally we ensure that exactly those values that satisfy the expected predicate are returned. We do this for both 'exact' and 'bounding_box' geometry.