Improve reliability of nearest neighbour test.
The test was unreliable because it expected that the spheres which are closest to the origin will be returned. Instead, Boost returns those spheres for which the bounding box is closest to the origin. To compensate for this it repeated the test.

This commit makes the test reflect what Boost does.