Use `RelWithDebInfo`.
The relevant difference between `Release` and `RelWithDebInfo`
is that the latter doesn't set `-DNDEBUG`. Therefore, `assert`s
are not optimized out.

The idea is to maximize our chances of noticing faults. That are
triggered by realistic circuits, but not yet covered by our
unit-tests.

Once we fully hit production we can revert to `Release`.