Implement `spatial_index.open_index`.
Also updates all examples and documentation. In order to allow multiple
incompatible versions of the data files to coexists, this commit introduces
a restructure of the data files used for CI and examples.

In order to facilitate these changes, every index is stored in a directory.
This directory contains the actual index and a meta data file. The directory
is expected to either not exist, or be empty. Unlike the current behaviour,
indexes are never overwritten.

Structure of the meta data file:
```
{
  "version": 0,
  "element_type": "morpho_entry",

  # Only for memory mapped indexes:
  "memory_mapped": {
    # anything memory mapped file related.
    "heavy_data_path": "index.bin"
  }

  # Only for in memory indexes:
  "in_memory": { ... }

  # Only for multi indexes:
  "multi_index": { ... }

  # Eventually for multi populations:
  "multi_population": {
    "populations": [
      {"population": "NodeA", "index_path": "node_a.spi"},
      ...
    ]
  }

  # Optionally:
  "extended": {
    # Probably things like path to the
    # SONATA data file, etc.
  }
}
```
The choice to go with JSON (the comments in the previous example are not legal) is because as far as I can tell JSON is the common choice for config files at BBP. Using TOML would require an additional C++ dependency on TOML. We could now drop the TOML dependency from SpatialIndex.

The design ideas for this structure are:
* Each section can set the setting it needs. This makes separate sections independent and decouples the individual section from each other.
* All paths in the meta data file are either relative to the directory which contains the meta data file; or absolute. If symlinks are involved then the directory is the path passed to us, not its `realpath`.