Index Structure V2 and BluePy validation
This MR switches the indexing for SpatialIndex to be `0-based` instead of `1-based`. This was explicitly requested to be compliant with future software such as `BluePySnap`. This is detailed in: https://bbpteam.epfl.ch/project/issues/browse/SPIND-176

It also integrates validation tests for SpatialIndex against BluePy to verify that the morphologies return the same results.
The tests performs 20 non-empty queries with expanding window, gets the results and then cross-checks that these results coincide with the values given by BluePy.
At the moment the tests performs "only" 20 queries to keep the execution time short enough for CI but the script can also be used independently to check as many random queries as needed.

The MR also fixes the overwriting bug that we had while analising incorrect results for the `section_id` field during the first implementation of this script. The `section_id` field is now 14 bits as it should fit all the circuits that we tested. It also aligns `SpatialIndex` with other software such as `TouchDetector` that allocate the same amount of bits for that field.