Removing Python bindings for Memory Mapped files
Since most of the functionalities of Memory Mapped files have been superseded by the Multi-Index, we are removing support for this method of indexing. This is done in order to simplify the functionalities of Spatial Index and to not confuse too much the user with a method that right now has some pretty limited use case (if any).

This MR removes all the Python structure and bindings related to the creation, loading and use of memory mapped files. The underlying C++ structure will be kept for the time being.

We also updated documentation, CI and examples to reflect this change.