Make most special queries redundant.
This commit unfortunately does two things. It makes most of our special query functions redundant and removes them. However, for this to work we need to properly support the index of spheres in the new API.

This step of making the special query functions redundant is to prepare for optimizing the dict-of-results query. The other reason for this change is to enable safely mixing all possible values of `fields` and ensure that only one query is run.