Make assert on radii length sharper.
We require that the there is one radius per point, meaning one for soma (if present) and for each section one more than the number of segments in that section. Essentially, we require that it has the same structure as `points`. Since SI only deals with cylinders we consider the "first" radius as "the" radius of the cylinder; and ignore the radius at the other cap.

This commit makes the asserts sharper and the error message clearer.