Radius of segment as average of the radius of the two endcaps
As discussed during the SpatialIndex update meeting of 21/09/2022 (with minutes available here: https://bbpteam.epfl.ch/project/spaces/x/cR2tBg ) we should consider the radius of a cylinder as the average of the two values supplied instead of discarding the second one like it happens now.

The geometry should stay linear and cylindrical.