Add a barrier after creating the meta data.
This should prevent race conditions, e.g. when an MPI rank that isn't the rank
writing the meta data file tries to open the index. If that rank is faster than
the rank writing the meta data, it may see no file or a partially written file,
without this barrier.