Fixes wheels and improves warning message
As reported in SPIND-217, the current wheel build of SpatialIndex is broken. This MR attempts to fix the issue by putting two guards that allow user to still use the program as long as they don't use any MPI-related that are unavailable in a wheel build anyway.

It also improves the warning message for users.