Improve error message when reading non-existant files.
This commit improves the message of the exception thrown when trying to open a non-existant file to state that the file doesn't exist and print the name of the file that should have been opened. Additionally, it also checks if the file can be read, and throw a slightly more meaningful message than the boost `input stream error` if it can't be opened for reading.