Enable `--verbose` from CLI to activate all log messages.
In order to set the loglevel we need to ensure that a root logger has been defined, see:
https://stackoverflow.com/a/43123658

This seems like a horrible idea when using SI as a library. Hence, this introduces a function that sets up the `logging` module, i.e. calls `logging.basicConfig()`, which is to be called in the CLI application we write for SI. Since those are our applications, it seems like we would be permitted to pick the logging behaviour, i.e. in that context we're allowed to call `logging.basicConfig` and configure the format to our liking.