Add uniform index generation script
This MR includes improvement and updates to the benchmarking scripts.
It includes also the `UniformFactory` class that allows to create synthetic indexes.
It also includes a script to generate uniform indexes, useful for benchmarking and tests.

A `README.md` for the `_benchmarking` folder was also added to explain how to use them and what to expect.